+++ src/c/utests/data/data.c
@@ -571,7 +571,7 @@ static void test_data_to_json (void)
   iot_data_free (val);
   free (json);
 
-  char buff[10];
+  char buff[32];
   val = iot_data_alloc_bool (true);
   json = iot_data_to_json_with_buffer (val, buff, sizeof (buff));
   CU_ASSERT (json == buff)
@@ -4033,7 +4033,18 @@ static void test_data_const_string (void)
   static const char * str = "Hello";
   iot_data_t * data = iot_data_alloc_const_string (&block, str);
   CU_ASSERT (iot_data_string (data) == str)
-  CU_ASSERT (data == IOT_DATA_STATIC (block))
+  CU_ASSERT (data == IOT_DATA_STATIC (&block))
+  CU_ASSERT (iot_data_is_static (data))
+  iot_data_free (data);
+  iot_data_free (data);
+}
+
+static void test_data_const_ui64 (void)
+{
+  static iot_data_static_t block;
+  iot_data_t * data = iot_data_alloc_const_ui64 (&block, 222u);
+  CU_ASSERT (iot_data_ui64 (data) == 222u)
+  CU_ASSERT (data == IOT_DATA_STATIC (&block))
   CU_ASSERT (iot_data_is_static (data))
   iot_data_free (data);
   iot_data_free (data);
@@ -4045,7 +4056,7 @@ static void test_data_const_pointer (void)
   static const void * ptr = &block;
   iot_data_t * data = iot_data_alloc_const_pointer (&block, ptr);
   CU_ASSERT (iot_data_pointer (data) == ptr)
-  CU_ASSERT (data == IOT_DATA_STATIC (block))
+  CU_ASSERT (data == IOT_DATA_STATIC (&block))
   CU_ASSERT (iot_data_is_static (data))
   iot_data_free (data);
   iot_data_free (data);
@@ -4061,7 +4072,7 @@ static void test_data_const_list (void)
   iot_data_t * str = iot_data_alloc_const_string (&block3, "test");
   CU_ASSERT (iot_data_list_length (list) == 0u)
   CU_ASSERT (iot_data_type (list) == IOT_DATA_LIST)
-  CU_ASSERT (list == IOT_DATA_STATIC (block))
+  CU_ASSERT (list == IOT_DATA_STATIC (&block))
   CU_ASSERT (iot_data_is_static (list))
   iot_data_list_tail_push (list, ptr);
   CU_ASSERT (iot_data_list_length (list) == 1u)
@@ -4076,7 +4087,7 @@ static void test_data_const_list (void)
   iot_data_t * list_non_static = iot_data_alloc_const_list (&block_non_static);
   CU_ASSERT (iot_data_list_length (list_non_static) == 0u)
   CU_ASSERT (iot_data_type (list_non_static) == IOT_DATA_LIST)
-  CU_ASSERT (list_non_static == IOT_DATA_STATIC (block_non_static))
+  CU_ASSERT (list_non_static == IOT_DATA_STATIC (&block_non_static))
   CU_ASSERT (iot_data_is_static (list_non_static))
   iot_data_list_tail_push (list_non_static, str);
   CU_ASSERT (iot_data_list_length (list_non_static) == 1u)
@@ -5160,6 +5171,7 @@ void cunit_data_test_init (void)
   CU_add_test (suite, "data_alloc_heap", test_data_alloc_heap);
   CU_add_test (suite, "data_cast", test_data_cast);
   CU_add_test (suite, "data_const_string", test_data_const_string);
+  CU_add_test (suite, "data_const_ui64", test_data_const_ui64);
   CU_add_test (suite, "data_const_pointer", test_data_const_pointer);
   CU_add_test (suite, "data_const_list", test_data_const_list);
   CU_add_test (suite, "data_const_types", test_data_const_types);