+++ src/c/utests/scheduler/scheduler.c
@@ -25,7 +25,7 @@ static void reset_counters (void)
   atomic_store (&counter, 0);
 }
 
-static void * do_work1 (void *in)
+static void * do_work1 (void * in)
 {
   (void) in;
   for (uint32_t i = 0; i < 10; ++i)
@@ -35,7 +35,7 @@ static void * do_work1 (void *in)
   return NULL;
 }
 
-static void * do_work2 (void *in)
+static void * do_work2 (void * in)
 {
   (void) in;
   for (uint32_t i = 0; i < 20; ++i)
@@ -45,7 +45,7 @@ static void * do_work2 (void *in)
   return NULL;
 }
 
-static void * do_work3 (void *in)
+static void * do_work3 (void * in)
 {
   (void) in;
   for (int i = 0; i < 30; ++i)
@@ -55,9 +55,17 @@ static void * do_work3 (void *in)
   return NULL;
 }
 
-static void * do_work4 (void *in)
+static void * do_work4 (void * in)
 {
   (void) in;
+  atomic_fetch_add (&sum_test, 1u);
+  return NULL;
+}
+
+static void * do_sum_100 (void * in)
+{
+  (void) in;
+  iot_wait_msecs (100u);
   atomic_fetch_add (&sum_test, 1u);
   return NULL;
 }
@@ -78,7 +86,7 @@ static void * do_count (void *in)
 
 static void * sched_create (void * data)
 {
-  *(int *)data = 10;
+  *(int*) data = 10;
   return data;
 }
 
@@ -402,10 +410,81 @@ static void cunit_scheduler_setfreefn (void)
   iot_scheduler_stop (scheduler);
   iot_threadpool_free (pool);
   iot_scheduler_free (scheduler);
-
 }
 
-extern void cunit_scheduler_test_init ()
+static void cunit_scheduler_rate (bool concurrent)
+{
+  reset_counters ();
+  iot_threadpool_t *pool = iot_threadpool_alloc (4u, 0u, IOT_THREAD_NO_PRIORITY, IOT_THREAD_NO_AFFINITY, logger);
+  iot_scheduler_t *scheduler = iot_scheduler_alloc (IOT_THREAD_NO_PRIORITY, IOT_THREAD_NO_AFFINITY, logger);
+  iot_schedule_t *sched = iot_schedule_create (scheduler, do_sum_100, NULL, NULL, IOT_MS_TO_NS (50u), 0, 0, pool, -1);
+  iot_schedule_set_concurrent (sched, concurrent);
+  CU_ASSERT (iot_schedule_add (scheduler, sched));
+  iot_threadpool_start (pool);
+  iot_scheduler_start (scheduler);
+  iot_wait_secs (2);
+  iot_scheduler_stop (scheduler);
+  iot_threadpool_free (pool);
+  iot_scheduler_free (scheduler);
+  uint32_t sum = atomic_load (&sum_test);
+  if (concurrent)
+  {
+    CU_ASSERT (sum > 21u)
+  }
+  else
+  {
+    CU_ASSERT (sum <= 21u)
+  }
+}
+
+static void cunit_scheduler_concurrent (void)
+{
+  cunit_scheduler_rate (true);
+}
+
+static void cunit_scheduler_serialized (void)
+{
+  cunit_scheduler_rate (false);
+}
+
+static void * scheduler_delete_with_user_data_do_work (void * arg)
+{
+  iot_wait_secs (2); //do some work
+  uint64_t *arg2 = arg;
+  (*arg2)++;
+  return NULL;
+}
+
+static void scheduler_delete_with_user_data_free (void *data)
+{
+  free (data);
+}
+
+static void cunit_scheduler_delete_with_user_data (void)
+{
+  uint64_t *arg = calloc (1, sizeof(*arg));
+  iot_scheduler_t * scheduler = iot_scheduler_alloc (IOT_THREAD_NO_PRIORITY, IOT_THREAD_NO_AFFINITY, NULL);
+  iot_schedule_t * sched1 = iot_schedule_create (
+    scheduler,
+    scheduler_delete_with_user_data_do_work,
+    scheduler_delete_with_user_data_free,
+    arg,
+    10,
+    0,
+    0,
+    NULL,
+    -1
+  );
+  CU_ASSERT_TRUE (iot_schedule_add (scheduler, sched1));
+  iot_scheduler_start (scheduler);
+  iot_wait_secs (1);
+  iot_schedule_delete (scheduler, sched1);
+  iot_wait_secs (2);
+  iot_scheduler_stop (scheduler);
+  iot_scheduler_free (scheduler);
+}
+
+extern void cunit_scheduler_test_init (void)
 {
   CU_pSuite suite = CU_add_suite ("scheduler", suite_init, suite_clean);
   CU_add_test (suite, "scheduler_start", cunit_scheduler_start);
@@ -419,5 +498,8 @@ extern void cunit_scheduler_test_init ()
   CU_add_test (suite, "scheduler_reset", cunit_scheduler_reset);
   CU_add_test (suite, "scheduler_setpriority", cunit_scheduler_setpriority);
   CU_add_test (suite, "scheduler_freefn", cunit_scheduler_setfreefn);
+  CU_add_test (suite, "scheduler_concurrent", cunit_scheduler_concurrent);
+  CU_add_test (suite, "scheduler_serialized", cunit_scheduler_serialized);
+  CU_add_test (suite, "scheduler_delete_with_user_data", cunit_scheduler_delete_with_user_data);
 }